🐛[FIX] speed glitch + 3d models

+++ Arduino/BikeeperFirmware/Headers/location.h
@@ -3,8 +3,8 @@
 
 struct location_t
 {
-	double latitude;
-	double longitude;
+	float latitude;
+	float longitude;
 
 	location_t() : 
 		latitude(),
